backend/requirements.txt
# Backend dependencies — NO Streamlit
openai>=1.50.0
openpyxl>=3.1.5
Pillow>=10.0.0
pdf2image>=1.17.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
# Sales Order module dependencies
pandas>=2.1.0
pymupdf>=1.23.0
pydantic>=2.0.0
psycopg2-binary
